fixes #31 confirm before delete

+++ src/main/webapp/WEB-INF/jsp/site/butter/edit-script.jsp
@@ -18,7 +18,9 @@
   $(function() {
     $("input[name='excerpt'][type='radio']").change(function(e) {
       $("input[name='excerpt'][type='text']").toggle();
-      $("input[name='excerpt'][type='text']").prop("disabled", function(i, v) { return !v; });
+      $("input[name='excerpt'][type='text']").prop("disabled", function(i, v) {
+        return !v;
+      });
     });
     function toggleFullscreen(simplemde) {
       if (simplemde.isFullscreenActive()) {
@@ -87,7 +89,10 @@
     $formEditButter.on("submit", function(event) {
       event.preventDefault();
       var data = $formEditButter.serializeObject();
-      if (data.excerpt == "" && !$("input[type='text'][name='excerpt']").prop('disabled')) {
+      if (
+        data.excerpt == "" &&
+        !$("input[type='text'][name='excerpt']").prop("disabled")
+      ) {
         alert("수정내용을 입력해주세요");
         $("input[type='text'][name='excerpt']").focus();
         return;
@@ -122,33 +127,35 @@
       });
     });
 
-    $("#btn-butter-delete").one("click", function(e) {
-      $.ajax({
-        headers: { "X-CSRF-TOKEN": "${_csrf.token}" },
-        url: "/ajax/butter/${butter.id}",
-        type: "DELETE",
-        contentType: "application/json",
-        dataType: "json",
-        success: function(data) {
-          alert(data.msg);
-          window.location.href = data.url;
-        },
-        error: function(error) {
-          if (error.status === 400) {
-            if (error.responseJSON.fieldErrors) {
-              var msg = error.responseJSON.fieldErrors
-                .map(function(item) {
-                  return item.fieldError;
-                })
-                .join("/n");
-              alert(msg);
-            } else alert(error.responseJSON.msg);
-          } else if (error.status === 403 || error.status === 401) {
-            alert("로그인이 필요합니다.");
-            window.location.href = "/login.do";
+    $("#btn-butter-delete").click(function(e) {
+      if (confirm("버터를 삭제하시겠습니까?")) {
+        $.ajax({
+          headers: { "X-CSRF-TOKEN": "${_csrf.token}" },
+          url: "/ajax/butter/${butter.id}",
+          type: "DELETE",
+          contentType: "application/json",
+          dataType: "json",
+          success: function(data) {
+            alert(data.msg);
+            window.location.href = data.url;
+          },
+          error: function(error) {
+            if (error.status === 400) {
+              if (error.responseJSON.fieldErrors) {
+                var msg = error.responseJSON.fieldErrors
+                  .map(function(item) {
+                    return item.fieldError;
+                  })
+                  .join("/n");
+                alert(msg);
+              } else alert(error.responseJSON.msg);
+            } else if (error.status === 403 || error.status === 401) {
+              alert("로그인이 필요합니다.");
+              window.location.href = "/login.do";
+            }
           }
-        }
-      });
+        });
+      }
     });
   });
 </script>
